First batch of moving impala ddl to base_sql
