+++ ibis/impala/ddl.py
@@ -13,66 +13,24 @@
 # limitations under the License.
 
 import json
-import re
 
 import ibis.expr.datatypes as dt
-import ibis.expr.schema as sch
-from ibis.backends.base_sql import quote_identifier, type_to_sql_string
-from ibis.backends.base_sqlalchemy.compiler import DDL, DML
-
-fully_qualified_re = re.compile(r"(.*)\.(?:`(.*)`|(.*))")
-
-
-def _is_fully_qualified(x):
-    return bool(fully_qualified_re.search(x))
-
-
-def _is_quoted(x):
-    regex = re.compile(r"(?:`(.*)`|(.*))")
-    quoted, _ = regex.match(x).groups()
-    return quoted is not None
+from ibis.backends.base_sql import type_to_sql_string
+from ibis.backends.base_sql.ddl import (
+    BaseDDL,
+    BaseQualifiedSQLStatement,
+    CreateTable,
+    CreateTableWithSchema,
+    DropObject,
+    DropTable,
+    _sanitize_format,
+    format_schema,
+)
+from ibis.backends.base_sqlalchemy.compiler import DML
 
 
-class ImpalaQualifiedSQLStatement:
-    def _get_scoped_name(self, obj_name, database):
-        if database:
-            scoped_name = '{}.`{}`'.format(database, obj_name)
-        else:
-            if not _is_fully_qualified(obj_name):
-                if _is_quoted(obj_name):
-                    return obj_name
-                else:
-                    return '`{}`'.format(obj_name)
-            else:
-                return obj_name
-        return scoped_name
-
-
-class ImpalaDDL(DDL, ImpalaQualifiedSQLStatement):
+class ImpalaDML(DML, BaseQualifiedSQLStatement):
     pass
-
-
-class ImpalaDML(DML, ImpalaQualifiedSQLStatement):
-    pass
-
-
-class CreateDDL(ImpalaDDL):
-    def _if_exists(self):
-        return 'IF NOT EXISTS ' if self.can_exist else ''
-
-
-_format_aliases = {'TEXT': 'TEXTFILE'}
-
-
-def _sanitize_format(format):
-    if format is None:
-        return
-    format = format.upper()
-    format = _format_aliases.get(format, format)
-    if format not in ('PARQUET', 'AVRO', 'TEXTFILE'):
-        raise ValueError('Invalid format: {!r}'.format(format))
-
-    return format
 
 
 def _serdeproperties(props):
@@ -91,129 +49,6 @@ def _format_properties(props):
         tokens.append("  '{}'='{}'".format(k, v))
 
     return '(\n{}\n)'.format(',\n'.join(tokens))
-
-
-class CreateTable(CreateDDL):
-
-    """
-
-    Parameters
-    ----------
-    partition :
-
-    """
-
-    def __init__(
-        self,
-        table_name,
-        database=None,
-        external=False,
-        format='parquet',
-        can_exist=False,
-        partition=None,
-        path=None,
-        tbl_properties=None,
-    ):
-        self.table_name = table_name
-        self.database = database
-        self.partition = partition
-        self.path = path
-        self.external = external
-        self.can_exist = can_exist
-        self.format = _sanitize_format(format)
-        self.tbl_properties = tbl_properties
-
-    @property
-    def _prefix(self):
-        if self.external:
-            return 'CREATE EXTERNAL TABLE'
-        else:
-            return 'CREATE TABLE'
-
-    def _create_line(self):
-        scoped_name = self._get_scoped_name(self.table_name, self.database)
-        return '{} {}{}'.format(self._prefix, self._if_exists(), scoped_name)
-
-    def _location(self):
-        return "LOCATION '{}'".format(self.path) if self.path else None
-
-    def _storage(self):
-        # By the time we're here, we have a valid format
-        return 'STORED AS {}'.format(self.format)
-
-    @property
-    def pieces(self):
-        yield self._create_line()
-        for piece in filter(None, self._pieces):
-            yield piece
-
-    def compile(self):
-        return '\n'.join(self.pieces)
-
-
-class CTAS(CreateTable):
-
-    """
-    Create Table As Select
-    """
-
-    def __init__(
-        self,
-        table_name,
-        select,
-        database=None,
-        external=False,
-        format='parquet',
-        can_exist=False,
-        path=None,
-        partition=None,
-    ):
-        super().__init__(
-            table_name,
-            database=database,
-            external=external,
-            format=format,
-            can_exist=can_exist,
-            path=path,
-            partition=partition,
-        )
-        self.select = select
-
-    @property
-    def _pieces(self):
-        yield self._partitioned_by()
-        yield self._storage()
-        yield self._location()
-        yield 'AS'
-        yield self.select.compile()
-
-    def _partitioned_by(self):
-        if self.partition is not None:
-            return 'PARTITIONED BY ({})'.format(
-                ', '.join(
-                    quote_identifier(expr._name) for expr in self.partition
-                )
-            )
-        return None
-
-
-class CreateView(CTAS):
-
-    """Create a view"""
-
-    def __init__(self, table_name, select, database=None, can_exist=False):
-        super().__init__(
-            table_name, select, database=database, can_exist=can_exist
-        )
-
-    @property
-    def _pieces(self):
-        yield 'AS'
-        yield self.select.compile()
-
-    @property
-    def _prefix(self):
-        return 'CREATE VIEW'
 
 
 class CreateTableParquet(CreateTable):
@@ -250,43 +85,6 @@ class CreateTableParquet(CreateTable):
             raise NotImplementedError
 
         yield self._storage()
-        yield self._location()
-
-
-class CreateTableWithSchema(CreateTable):
-    def __init__(self, table_name, schema, table_format=None, **kwargs):
-        super().__init__(table_name, **kwargs)
-        self.schema = schema
-        self.table_format = table_format
-
-    @property
-    def _pieces(self):
-        if self.partition is not None:
-            main_schema = self.schema
-            part_schema = self.partition
-            if not isinstance(part_schema, sch.Schema):
-                part_schema = sch.Schema(
-                    part_schema, [self.schema[name] for name in part_schema]
-                )
-
-            to_delete = []
-            for name in self.partition:
-                if name in self.schema:
-                    to_delete.append(name)
-
-            if len(to_delete):
-                main_schema = main_schema.delete(to_delete)
-
-            yield format_schema(main_schema)
-            yield 'PARTITIONED BY {}'.format(format_schema(part_schema))
-        else:
-            yield format_schema(self.schema)
-
-        if self.table_format is not None:
-            yield '\n'.join(self.table_format.to_ddl())
-        else:
-            yield self._storage()
-
         yield self._location()
 
 
@@ -451,7 +249,7 @@ def _format_partition_kv(k, v, type):
     return '{}={}'.format(k, value_formatted)
 
 
-class LoadData(ImpalaDDL):
+class LoadData(BaseDDL):
 
     """
     Generate DDL for LOAD DATA command. Cannot be cancelled
@@ -491,7 +289,7 @@ class LoadData(ImpalaDDL):
         )
 
 
-class AlterTable(ImpalaDDL):
+class AlterTable(BaseDDL):
     def __init__(
         self,
         table,
@@ -615,30 +413,7 @@ class RenameTable(AlterTable):
         return self._wrap_command(cmd)
 
 
-class DropObject(ImpalaDDL):
-    def __init__(self, must_exist=True):
-        self.must_exist = must_exist
-
-    def compile(self):
-        if_exists = '' if self.must_exist else 'IF EXISTS '
-        object_name = self._object_name()
-        return 'DROP {} {}{}'.format(self._object_type, if_exists, object_name)
-
-
-class DropTable(DropObject):
-
-    _object_type = 'TABLE'
-
-    def __init__(self, table_name, database=None, must_exist=True):
-        super().__init__(must_exist=must_exist)
-        self.table_name = table_name
-        self.database = database
-
-    def _object_name(self):
-        return self._get_scoped_name(self.table_name, self.database)
-
-
-class TruncateTable(ImpalaDDL):
+class TruncateTable(BaseDDL):
 
     _object_type = 'TABLE'
 
@@ -656,7 +431,7 @@ class DropView(DropTable):
     _object_type = 'VIEW'
 
 
-class CacheTable(ImpalaDDL):
+class CacheTable(BaseDDL):
     def __init__(self, table_name, database=None, pool='default'):
         self.table_name = table_name
         self.database = database
@@ -669,50 +444,7 @@ class CacheTable(ImpalaDDL):
         )
 
 
-class CreateDatabase(CreateDDL):
-    def __init__(self, name, path=None, can_exist=False):
-        self.name = name
-        self.path = path
-        self.can_exist = can_exist
-
-    def compile(self):
-        name = quote_identifier(self.name)
-
-        create_decl = 'CREATE DATABASE'
-        create_line = '{} {}{}'.format(create_decl, self._if_exists(), name)
-        if self.path is not None:
-            create_line += "\nLOCATION '{}'".format(self.path)
-
-        return create_line
-
-
-class DropDatabase(DropObject):
-
-    _object_type = 'DATABASE'
-
-    def __init__(self, name, must_exist=True):
-        super().__init__(must_exist=must_exist)
-        self.name = name
-
-    def _object_name(self):
-        return self.name
-
-
-def format_schema(schema):
-    elements = [
-        _format_schema_element(name, t)
-        for name, t in zip(schema.names, schema.types)
-    ]
-    return '({})'.format(',\n '.join(elements))
-
-
-def _format_schema_element(name, t):
-    return '{} {}'.format(
-        quote_identifier(name, force=True), type_to_sql_string(t),
-    )
-
-
-class CreateFunction(ImpalaDDL):
+class CreateFunction(BaseDDL):
 
     _object_type = 'FUNCTION'
 
@@ -792,7 +524,7 @@ class DropFunction(DropObject):
         return ' '.join(tokens)
 
 
-class ListFunction(ImpalaDDL):
+class ListFunction(BaseDDL):
     def __init__(self, database, like=None, aggregate=False):
         self.database = database
         self.like = like
